scripts: Move json i/o functions to utils
The release tagging script already has a duplicate copy of
readPackageJson and it's going to soon need a copy of writePackageJson
as well.

Rather than copy and paste, this commit moves those functions from
update-package-json.js to the shared utils.js.
